requirements.txt
urllib3==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.4
sqlalchemy==2.0.15
geoalchemy2==0.13.3
shapely==2.0.1
fiona==1.9.4
pyproj==3.3.1
pandas==2.0.2
geopandas==0.13.2
rasterio==1.3.7
netCDF4==1.6.4
xlrd==2.0.1
xlwt==1.3.0
xlsxwriter==3.1.2
openpyxl==3.1.2
dbf==0.99.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
flickrapi==2.4.0
tweepy==4.10.0
scipy==1.10.1
scikit-learn==1.2.2
scikit-image==0.21.0
pyexcel-ods==0.6.0
bs4
seaborn==0.11.2
sentinelsat==1.2.1
simpledbf==0.2.6
#mysqlclient==1.4.6
xmltodict==0.13.0
jupyterlab
polyline==1.4.0
py_expression_eval==0.3.14
jenkspy==0.2.0
pysal==2.6.0
geopy==2.3.0
#rpy2==3.5.10
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-oauth-toolkit==2.1.0
drf-yasg==1.21.4
django-dbbackup==4.0.2